Refactor ft_swap function: change parameter types from double to float for consistency

+++ lib/libft/include/libft.h
@@ -6,7 +6,7 @@
 /*   By: lkilpela <lkilpela@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/26 14:57:48 by lkilpela          #+#    #+#             */
-/*   Updated: 2024/09/10 09:46:50 by lkilpela         ###   ########.fr       */
+/*   Updated: 2024/09/25 18:43:28 by lkilpela         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ int		ft_isascii(int c);
 int		ft_isalnum(int c);
 int		ft_isprint(int c);
 int		ft_isspace(int c);
-void	ft_swap(double *a, double *b);
+void	ft_swap(float *a, float *b);
 size_t	ft_strlen(const char *s);
 void	*ft_memset(void *b, int c, size_t len);
 void	ft_bzero(void *s, size_t n);
